move CW & Bank -> a modules mod
