+++ local-interchain/rust/localic-std/src/modules/cosmwasm.rs
@@ -28,8 +28,6 @@ impl CosmWasm<'_> {
         }
     }
 
-    /// new_from_existing is used when we want to use an existing code_id and contract_addr
-    /// but still use the CosmWasm methods to execute queries and transactions against it.
     pub fn new_from_existing(
         rb: &ChainRequestBuilder,
         file_path: Option<PathBuf>,
@@ -69,66 +67,12 @@ impl CosmWasm<'_> {
             None => panic!("contract_addr is none"),
         };
 
-        match self.contract_instantiate(account_key, code_id, msg, label, admin, flags) {
+        match contract_instantiate(self.rb, account_key, code_id, msg, label, admin, flags) {
             Ok(contract) => {
                 self.contract_addr = Some(contract.address.to_owned());
                 Ok(contract)
             }
             Err(e) => Err(e),
-        }
-    }
-
-    pub fn contract_instantiate(
-        &self,
-        account_key: &str,
-        code_id: u64,
-        msg: &str,
-        label: &str,
-        admin: Option<&str>,
-        flags: &str,
-    ) -> Result<Contract, LocalError> {
-        let mut updated_flags = flags.to_string();
-        if admin.is_none() && !flags.contains("--no-admin") {
-            updated_flags = format!("{} --no-admin", flags);
-        } else if admin.is_some() {
-            updated_flags = format!("{} --admin={}", flags, admin.unwrap());
-        }
-        updated_flags = updated_flags.trim().to_string();
-
-        let mut cmd = format!("tx wasm instantiate {code_id} {msg} --label={label} --from={account_key} --keyring-backend=test --node=%RPC% --chain-id=%CHAIN_ID% --output=json --gas=auto --gas-adjustment=3.0 --yes", code_id=code_id, msg=msg, label=label, account_key=account_key);
-        if !updated_flags.is_empty() {
-            cmd = format!("{} {}", cmd, updated_flags);
-        }
-
-        let res = self.rb.tx(cmd.as_str(), false);
-        if res.is_err() {
-            return Err(res.err().unwrap());
-        }
-
-        let res = res.unwrap();
-
-        println!("wasm instantiate res: {}", &res);
-
-        let tx_hash: Option<String> = (&self.rb).get_tx_hash(&res);
-        let raw_log: Option<String> = (&self.rb).get_raw_log(&res);
-
-        if raw_log.is_some() {
-            println!("raw_log: {}", raw_log.clone().unwrap());
-        }
-
-        let contract_addr =
-            get_contract_address(&self.rb, tx_hash.clone().unwrap_or_default().as_str());
-        match contract_addr {
-            Ok(contract_addr) => {
-                return Ok(Contract {
-                    address: contract_addr,
-                    tx_hash: tx_hash.unwrap_or_default(),
-                    admin: admin.map(|s| s.to_string()),
-                });
-            }
-            Err(e) => {
-                return Err(e);
-            }
         }
     }
 
@@ -142,65 +86,18 @@ impl CosmWasm<'_> {
             Some(addr) => addr.as_ref(),
             None => panic!("contract_addr is none"),
         };
-        self.execute_contract(contract_addr, account_key, msg, flags)
+        contract_execute(self.rb, contract_addr, account_key, msg, flags)
     }
 
-    pub fn execute_contract(
-        &self,
-        contract_addr: &str,
-        account_key: &str,
-        msg: &str,
-        flags: &str,
-    ) -> Result<TransactionResponse, LocalError> {
-        let mut cmd = format!(
-            "tx wasm execute {contract_addr} {msg} --from={account_key} --keyring-backend=test --home=%HOME% --node=%RPC% --chain-id=%CHAIN_ID% --yes {flags}",
-            contract_addr = contract_addr,
-            msg = msg,
-            account_key = account_key,
-            flags = flags
-        );
-        cmd = cmd.trim().to_string();
-
-        let updated_flags = flags.to_string();
-        if !updated_flags.is_empty() {
-            cmd = format!("{} {}", cmd, updated_flags);
-        }
-
-        let res = self.rb.binary(cmd.as_str(), false);
-        println!("execute_contract res: {}", &res);
-
-        let tx_hash = self.rb.get_tx_hash(&res);
-        let tx_raw_log = self.rb.get_raw_log(&res);
-
-        if let Some(raw_log) = &tx_raw_log {
-            if raw_log != "[]" {
-                println!("execute_contract raw_log: {}", raw_log);
-            }
-        }
-
-        Ok(TransactionResponse {
-            tx_hash,
-            rawlog: tx_raw_log,
-        })
-    }
-
+    
     pub fn query(&self, msg: &str) -> Value {
         let contract_addr: &str = match &self.contract_addr {
             Some(addr) => addr.as_ref(),
             None => panic!("contract_addr is none"),
         };
-        self.query_contract(contract_addr, msg)
+        contract_query(self.rb, contract_addr, msg)
     }
-
-    pub fn query_contract(&self, contract_addr: &str, msg: &str) -> Value {
-        let cmd = format!(
-            "query wasm contract-state smart {} {} --output=json --node=%RPC%",
-            contract_addr, msg
-        );
-        println!("query_contract cmd: {}", cmd);
-        let res = self.rb.query(&cmd, false);
-        res
-    }
+    
 
     pub fn create_wasm_connection(
         &self,
@@ -230,6 +127,109 @@ impl CosmWasm<'_> {
     }
 }
 
+pub fn contract_instantiate(
+    rb: &ChainRequestBuilder,
+    account_key: &str,
+    code_id: u64,
+    msg: &str,
+    label: &str,
+    admin: Option<&str>,
+    flags: &str,
+) -> Result<Contract, LocalError> {
+    let mut updated_flags = flags.to_string();
+    if admin.is_none() && !flags.contains("--no-admin") {
+        updated_flags = format!("{} --no-admin", flags);
+    } else if admin.is_some() {
+        updated_flags = format!("{} --admin={}", flags, admin.unwrap());
+    }
+    updated_flags = updated_flags.trim().to_string();
+
+    let mut cmd = format!("tx wasm instantiate {code_id} {msg} --label={label} --from={account_key} --keyring-backend=test --node=%RPC% --chain-id=%CHAIN_ID% --output=json --gas=auto --gas-adjustment=3.0 --yes");
+    if !updated_flags.is_empty() {
+        cmd = format!("{} {}", cmd, updated_flags);
+    }
+
+    let res = rb.tx(cmd.as_str(), false);
+    if res.is_err() {
+        return Err(res.err().unwrap());
+    }
+
+    let res = res.unwrap();
+
+    println!("wasm instantiate res: {}", &res);
+
+    let tx_hash: Option<String> = rb.get_tx_hash(&res);
+    let raw_log: Option<String> = rb.get_raw_log(&res);
+
+    if raw_log.is_some() {
+        println!("raw_log: {}", raw_log.clone().unwrap());
+    }
+
+    let contract_addr =
+        get_contract_address(rb, tx_hash.clone().unwrap_or_default().as_str());
+    match contract_addr {
+        Ok(contract_addr) => {
+            return Ok(Contract {
+                address: contract_addr,
+                tx_hash: tx_hash.unwrap_or_default(),
+                admin: admin.map(|s| s.to_string()),
+            });
+        }
+        Err(e) => {
+            return Err(e);
+        }
+    }
+}
+
+pub fn contract_execute(
+    rb: &ChainRequestBuilder,
+    contract_addr: &str,
+    account_key: &str,
+    msg: &str,
+    flags: &str,
+) -> Result<TransactionResponse, LocalError> {
+    let mut cmd = format!(
+        "tx wasm execute {contract_addr} {msg} --from={account_key} --keyring-backend=test --home=%HOME% --node=%RPC% --chain-id=%CHAIN_ID% --yes {flags}",
+        contract_addr = contract_addr,
+        msg = msg,
+        account_key = account_key,
+        flags = flags
+    );
+    cmd = cmd.trim().to_string();
+
+    let updated_flags = flags.to_string();
+    if !updated_flags.is_empty() {
+        cmd = format!("{} {}", cmd, updated_flags);
+    }
+
+    let res = rb.binary(cmd.as_str(), false);
+    println!("execute_contract res: {}", &res);
+
+    let tx_hash = rb.get_tx_hash(&res);
+    let tx_raw_log = rb.get_raw_log(&res);
+
+    if let Some(raw_log) = &tx_raw_log {
+        if raw_log != "[]" {
+            println!("execute_contract raw_log: {}", raw_log);
+        }
+    }
+
+    Ok(TransactionResponse {
+        tx_hash,
+        rawlog: tx_raw_log,
+    })
+}
+
+pub fn contract_query(rb: &ChainRequestBuilder, contract_addr: &str, msg: &str) -> Value {
+    let cmd = format!(
+        "query wasm contract-state smart {} {} --output=json --node=%RPC%",
+        contract_addr, msg
+    );
+    println!("query_contract cmd: {}", cmd);
+    let res = rb.query(&cmd, false);
+    res
+}
+
 pub fn get_contract_address(rb: &ChainRequestBuilder, tx_hash: &str) -> Result<String, LocalError> {
     let mut last_error = LocalError::ContractAddressNotFound {
         events: "".to_string(),
@@ -248,7 +248,7 @@ pub fn get_contract_address(rb: &ChainRequestBuilder, tx_hash: &str) -> Result<S
 }
 
 fn get_contract(rb: &ChainRequestBuilder, tx_hash: &str) -> Result<String, LocalError> {
-    let res = rb.query_tx_hash(tx_hash.to_string());
+    let res = rb.query_tx_hash(tx_hash);
 
     let code = res["code"].as_i64().unwrap_or_default();
     if code != 0 {
