+++ local-interchain/rust/localic-std/src/modules/mod.rs
@@ -1,0 +1,2 @@
+pub mod cosmwasm;
+pub mod bank;
